Changed Float to Double

+++ src/model/PathState.scala
@@ -4,7 +4,7 @@ package model
   * Created by agieg on 4/18/2017.
   */
 class PathState extends State {
-  def successors(): List[(State,Float)] = {
+  def successors(): List[(State,Double)] = {
     List()
   }
 }
